+++ src/client/testing/testController/common/server.ts
@@ -24,6 +24,10 @@ export class PythonTestServer implements ITestServer, Disposable {
 
     private ready: Promise<void>;
 
+    private _onRunDataReceived: EventEmitter<DataReceivedEvent> = new EventEmitter<DataReceivedEvent>();
+
+    private _onDiscoveryDataReceived: EventEmitter<DataReceivedEvent> = new EventEmitter<DataReceivedEvent>();
+
     constructor(private executionFactory: IPythonExecutionFactory, private debugLauncher: ITestDebugLauncher) {
         this.server = net.createServer((socket: net.Socket) => {
             let buffer: Buffer = Buffer.alloc(0); // Buffer to accumulate received data
@@ -48,11 +52,28 @@ export class PythonTestServer implements ITestServer, Disposable {
                         rawData = rpcHeaders.remainingRawData;
                         const rpcContent = jsonRPCContent(rpcHeaders.headers, rawData);
                         const extractedData = rpcContent.extractedJSON;
+                        // do not send until we have the full content
                         if (extractedData.length === Number(totalContentLength)) {
-                            // do not send until we have the full content
-                            traceVerbose(`Received data from test server: ${extractedData}`);
-                            this._onDataReceived.fire({ uuid, data: extractedData });
-                            this.uuids = this.uuids.filter((u) => u !== uuid);
+                            // if the rawData includes tests then this is a discovery request
+                            if (rawData.includes(`"tests":`)) {
+                                this._onDiscoveryDataReceived.fire({
+                                    uuid,
+                                    data: rpcContent.extractedJSON,
+                                });
+                                // if the rawData includes result then this is a run request
+                            } else if (rawData.includes(`"result":`)) {
+                                this._onRunDataReceived.fire({
+                                    uuid,
+                                    data: rpcContent.extractedJSON,
+                                });
+                            } else {
+                                traceLog(
+                                    `Error processing test server request: request is not recognized as discovery or run.`,
+                                );
+                                this._onDataReceived.fire({ uuid: '', data: '' });
+                                return;
+                            }
+                            // this.uuids = this.uuids.filter((u) => u !== uuid); WHERE DOES THIS GO??
                             buffer = Buffer.alloc(0);
                         } else {
                             break;
@@ -95,6 +116,18 @@ export class PythonTestServer implements ITestServer, Disposable {
         const uuid = crypto.randomUUID();
         this.uuids.push(uuid);
         return uuid;
+    }
+
+    public deleteUUID(uuid: string): void {
+        this.uuids = this.uuids.filter((u) => u !== uuid);
+    }
+
+    public get onRunDataReceived(): Event<DataReceivedEvent> {
+        return this._onRunDataReceived.event;
+    }
+
+    public get onDiscoveryDataReceived(): Event<DataReceivedEvent> {
+        return this._onDiscoveryDataReceived.event;
     }
 
     public dispose(): void {
